reimplementation (#44)
* moved old stuff

* Update README.md

* fixed

* data types

* reimplemented types

* added types

* updated

* slot and validator stuff

* committee stuff

* crosslink committees at slot

* more helper functions

* more helpers

* vote stuff

* done with "basic" helpers

* commented some tests out

* removed

* wrapper func

* initialization

* validator processing

* test fixes

* slot processing

* proposerSlashing

* calling func

* randao

* eth1data

* proposer slashings

* casper slashings

* actually calling

* attestations done

* deposits

* did exit

* fixes

* defined epoch helpers

* added some stuff

* cleanup

* updated

* Validators targeting the previous justified slot

* added more stuffs

* added in helpers

* using total balance

* inclusionSlot and distance

* justification call

* expected ffg source

* ffg target

* expected beaconchain head

* attestation inclusion

* crosslink

* process ejections

* updates shuffling seed data

* added if block for registry update

* update registry

* processPenaltiesAndExit

* test fix

* fix

* addeda  todo

* changed

+++ Sources/BeaconChain/DataStructures/Operations/Deposits/Deposit.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 struct Deposit {
-    let branch: [Data] // not sure if good
-    let index: Int
+    let branch: [Data]
+    let index: UInt64
     let depositData: DepositData
 }
