+++ Sources/BeaconChain/DataStructures/State/Eth1DataVote.swift
@@ -2,6 +2,5 @@ import Foundation
 
 struct Eth1DataVote {
     let eth1Data: Eth1Data
-    var voteCount: Int
-
+    var voteCount: UInt64
 }
